4.1.1-v1.0.0-Pre1

+++ utilities.py
@@ -16,7 +16,8 @@ def is_admin():
 
 def backup_file(src, backup_suffix=".backup"):
     backup_path = src + backup_suffix
-    shutil.copy2(src, backup_path)
+    if not os.path.exists(backup_path):
+        shutil.copy2(src, backup_path)
     return backup_path
 
 def replace_file(src, dst):
@@ -55,7 +56,8 @@ async def wait_for_exit(proc :str):
     raise ValueError(f"No process found with name containing: {proc}")
 
 def is_process_running(name_substring):
-    any(name_substring.lower() in proc.info['name'].lower() for proc in psutil.process_iter(['pid', 'name']))
+    print(f"Checking if process containing: {name_substring} is running")
+    return any(name_substring.lower() in proc.info['name'].lower() for proc in psutil.process_iter(['pid', 'name']))
 
 def kill_process_by_name(name_substring):
     for proc in psutil.process_iter(['pid', 'name']):
